pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "payment-router"
version = "0.1.0"
description = "Payment provider routing simulator — realistic approval rates, decline codes, latency, 3DS, and webhooks for integration testing"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Ivan Antonov", email = "ivan.antonov@outlook.com" }]
keywords = ["payments", "routing", "simulator", "fintech", "testing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Office/Business :: Financial",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.1",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "numpy>=1.26",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "redis>=5.0",
    "celery>=5.4",
    "httpx>=0.27",
    "structlog>=24.1",
    "kafka-python>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
]

[project.scripts]
payment-router = "payment_router.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["payment_router*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
